[core] A2解决bm/trans共同拉起失败的问题



# message auto-generated for no-merge-commit merge:
!820 merge release1.1_bug_fix into release/1.1

[core] A2解决bm/trans共同拉起失败的问题

Created-by: mrh1024
Commit-by: mrh1024
Merged-by: chenyz6
Description: # Pull Request Template

## Description

<!-- Provide a clear summary of the change, the problem it solves, and the technical approach. -->

## Related Issues

<!-- Replace with actual issue numbers. Example: Closes #123, Related to #456 -->

## Type of Change

- [ ] Bug fix
- [ ] New feature
- [ ] Performance optimization
- [ ] Documentation update
- [ ] Other (please describe):_______________

## Testing

<!-- Describe how the changes were tested. Include:
- Unit/integration test coverage
- Manual validation steps
- Links to CI reports (e.g., CIDA) if available
- Note any pending tests
-->

## Checklist

- [ ] I have performed a self-review of my own code.  
- [ ] I have updated the documentation.  
- [ ] I have added tests to prove my changes are effective.

See merge request: Ascend/memfabric_hybrid!820

+++ src/hybm/csrc/under_api/dl_hccp_def.h
@@ -153,7 +153,7 @@ enum HccpNotifyType {
 struct HccpSocketConnectInfo {
     void *handle;                      /**< socket handle */
     HccpIpAddr remoteIp;               /**< IP address of remote socket, [0-7] is reserved for vnic */
-    uint16_t port;                     /**< Socket listening port number */
+    uint32_t port;                     /**< Socket listening port number */
     char tag[HCCP_SOCK_CONN_TAG_SIZE]; /**< tag must ended by '\0' */
 };
